+++ CHANGELOG.md
@@ -426,3 +426,4 @@
 - 2024-06-08: Async mutate, mailbox, qualify, mod, +3
 - 2024-06-09: Health user, send, mutate,  outreach base, +4
 - 2024-06-10: Init leads, mailbox, bettercontact, descend, +2
+- 2024-06-11: Init follow up agent, contacts, mutate, leads, +3